feat : add payment function
결제, 리다이렉트, Status 변경 설정 완료

결제 검증로직추가

+++ src/main/java/org/hanghae/markethub/domain/purchase/dto/PurchaseRequestDto.java
@@ -3,19 +3,21 @@ package org.hanghae.markethub.domain.purchase.dto;
 import lombok.Builder;
 import org.hanghae.markethub.global.constant.Status;
 
+import java.math.BigDecimal;
+
 public record PurchaseRequestDto(
         Status status,
         String itemName,
         String email,
         int quantity,
-        int price
+        BigDecimal price
 ) {
     public record SinglePurchaseRequestDto(
             Status status,
             String itemName,
             String email,
             int quantity,
-            int price
+            BigDecimal price
     ) {
 
     }
