+++ src/main/java/org/hanghae/markethub/domain/purchase/dto/PurchaseResponseDto.java
@@ -4,6 +4,7 @@ import org.hanghae.markethub.domain.purchase.entity.Purchase;
 import org.hanghae.markethub.global.constant.Status;
 
 
+import java.math.BigDecimal;
 import java.util.List;
 import java.util.stream.Collectors;
 
@@ -32,7 +33,7 @@ public record PurchaseResponseDto(
 
     public record ItemDetailsDto(
             String itemName,
-            int price,
+            BigDecimal price,
             int quantity
     ) {
     }
